Revert "Feature: NavigationBar seperation from app.tsx"
