--- mobile/docs/dev_journal/Ame.md
@@ -1,8 +1,0 @@
-# downloading npm
-if you want to use simple commadns first download expo:npm install expo
-
-# resolving issue opening ios emulator
-For the first time opening this project the ios emulator is going to be behind after you add changes for the web app. 
-You have to do command: npx expo start -c
-And then refresh like 3 times
-Main question do we have to do it multiple times?
